增加链表存储帧数据

+++ client/pojo/Packet.h
@@ -1,7 +1,9 @@
 //
 // Created by 24546 on 2023/9/26.
 //
-
+#include <pthread.h>
+#include <stdbool.h>
+#include <stdlib.h>
 #ifndef STREAM_CLIENT_PACKET_H
 #define STREAM_CLIENT_PACKET_H
 
@@ -12,7 +14,7 @@ struct Packet{
     //服务器发送这个包的时间戳
     int sendTime;
     //客户端接收到这个包的时间戳
-    int revTime;
+    long long revTime;
 
     struct Packet *next;
 };
